optimize Kaleidoscope
