--- www/src/components/Canvas/index.js
@@ -1,1 +1,0 @@
-export * from "./Canvas"
