+++ www/src/pages/index.jsx
@@ -19,8 +19,6 @@ import Project from "../views/Project"
 import FeaturedProject from "../views/FeaturedProject"
 import VectorInverseStar from "../images/VectorInverseStar"
 import { colors } from "../styles/variables"
-import TestHero from "../views/TestHero"
-import TestMemoHero from "../views/TestMemoHero"
 
 const IndexPage = ({ data }) => (
   <Layout>
@@ -40,11 +38,7 @@ const IndexPage = ({ data }) => (
       ]}
     />
 
-    {/* <Hero /> */}
-
-    {/* <TestHero /> */}
-
-    <TestMemoHero />
+    <Hero />
 
     <Home />
 
@@ -66,7 +60,7 @@ const IndexPage = ({ data }) => (
     >
       <source
         sizes="100vw"
-        srcset="
+        srcSet="
           /images/bg-mandats_qo3wfk_c_scale,w_320.webp 320w,
           /images/bg-mandats_qo3wfk_c_scale,w_1661.webp 1661w,
           /images/bg-mandats_qo3wfk_c_scale,w_2170.webp 2170w,
@@ -89,7 +83,7 @@ const IndexPage = ({ data }) => (
 
       <img
         sizes="100vw"
-        srcset="
+        srcSet="
         /images/bg-mandats_qvp8kl_c_scale,w_320.jpg 320w,
         /images/bg-mandats_qvp8kl_c_scale,w_1128.jpg 1128w,
         /images/bg-mandats_qvp8kl_c_scale,w_1467.jpg 1467w,
@@ -154,7 +148,7 @@ const IndexPage = ({ data }) => (
       >
         <source
           sizes="100vw"
-          srcset="
+          srcSet="
         /images/bg-organisation,w_256.webp 256w,
         /images/bg-organisation,w_640.webp 640w,
         /images/bg-organisation,w_844.webp 844w,
@@ -173,7 +167,7 @@ const IndexPage = ({ data }) => (
           alt=""
           role="presentation"
           sizes="100vw"
-          srcset="
+          srcSet="
         /images/bg-organisation,w_256.jpg 256w,
         /images/bg-organisation,w_640.jpg 640w,
         /images/bg-organisation,w_844.jpg 844w,
